requirements.txt
streamlit
langchain
pypdf
openai
faiss-cpu
tiktoken
python-dotenv
